+++ client/components/Admin/index.js
@@ -1,3 +1,4 @@
-export { default as Settings } from './Settings';
+export { default as Applicants } from './Applicants';
 export { default as Landing } from './Landing';
 export { default as Review } from './Review';
+export { default as Settings } from './Settings';
